Lab5: Added useFetchLogger hook and refactored MenuPage structure
