+++ src/App.jsx
@@ -1,13 +1,15 @@
 import { useState } from "react";
 import HomePage from "./pages/HomePage/HomePage";
-import MenuPage from './components/MenuPage/MenuPage';
+import MenuPage from "./pages/MenuPage/MenuPage";
 
 function App() {
   const [activePage, setActivePage] = useState("home");
 
   return (
     <>
-      {activePage === "home" && <HomePage onNavigate={() => setActivePage("menu")} />}
+      {activePage === "home" && (
+        <HomePage onNavigate={() => setActivePage("menu")} />
+      )}
       {activePage === "menu" && <MenuPage />}
     </>
   );
